+++ pkg/config/config.go
@@ -1,40 +1,38 @@
 package config
 
 type Config struct {
-	Network  NetworkConfig   `mapstructure:"network"`
-	Content  ContentConfig   `mapstructure:"content"`
-	Users    map[string]User `mapstructure:"users"`
-	Security SecurityConfig  `mapstructure:"security"`
+	Network  NetworkConfig   `yaml:"network"`
+	Content  ContentConfig   `yaml:"content"`
+	Users    map[string]User `yaml:"users"`
+	Security SecurityConfig  `yaml:"security"`
 }
 
 type SecurityConfig struct {
-	AuthType string `mapstructure:"authtype"`
+	AuthType string `yaml:"authtype"`
 }
 
 type NetworkConfig struct {
-	Address string `mapstructure:"address"`
-	Port    string `mapstructure:"port"`
-	Prefix  string `mapstructure:"prefix,omitempty"`
+	Address string `yaml:"address"`
+	Port    string `yaml:"port"`
 }
 
 type ContentConfig struct {
-	Dir            string   `mapstructure:"dir"`
-	SubDirectories []string `mapstructure:"subdirectories,omitempty"`
+	Dir            string   `yaml:"dir"`
+	SubDirectories []string `yaml:"subdirectories,omitempty"`
 }
 
 type User struct {
-	Password       string   `mapstructure:"password"`
-	Root           string   `mapstructure:"root,omitempty"`
-	SubDirectories []string `mapstructure:"subdirectories,omitempty"`
-	Jail           bool     `mapstructure:"jail,omitempty"`
-	Admin          bool     `mapstructure:"admin"`
+	Password       string   `yaml:"password"`
+	Root           string   `yaml:"root,omitempty"`
+	SubDirectories []string `yaml:"subdirectories,omitempty"`
+	Jail           bool     `yaml:"jail,omitempty"`
+	Admin          bool     `yaml:"admin"`
 }
 
 var configTemplate = Config{
 	Network: NetworkConfig{
 		Address: "0.0.0.0",
 		Port:    "8080",
-		Prefix:  "/",
 	},
 	Content: ContentConfig{
 		Dir:            "/var/webdav/data",
@@ -51,7 +49,6 @@ func DeepCopyConfig(original Config) Config {
 		Network: NetworkConfig{
 			Address: original.Network.Address,
 			Port:    original.Network.Port,
-			Prefix:  original.Network.Prefix,
 		},
 		Security: SecurityConfig{
 			AuthType: original.Security.AuthType,
@@ -66,7 +63,7 @@ func DeepCopyConfig(original Config) Config {
 }
 
 type EnvironmentConfig struct {
-	WebdavPort        string
-	CreateNoAdminUser bool
-	WebdavDataDir     string
+	WebdavPort    string
+	WebdavDataDir string
+	AuthType      string
 }