+++ CHANGELOG.md
@@ -1,17 +1,38 @@
-# [3.1.0](https://github.com/triargos/webdav/compare/v3.0.0...v3.1.0) (2024-07-17)
+# [3.2.0-dev.2](https://github.com/triargos/webdav/compare/v3.2.0-dev.1...v3.2.0-dev.2) (2024-07-25)
+
+
+### Bug Fixes
+
+* **docs:** remove unused prefix from docs ([f958994](https://github.com/triargos/webdav/commit/f95899456712322321725b2bdf238cd0bd837340))
+
+# [3.2.0-dev.1](https://github.com/triargos/webdav/compare/v3.1.1-dev.1...v3.2.0-dev.1) (2024-07-25)
+
+
+### Bug Fixes
+
+* **cmd:** name resetConfigCmd correctly ([f652dda](https://github.com/triargos/webdav/commit/f652dda582dd084a0230092521bc6d1c744bb83e))
+
+
+### Features
+
+* **cmd:** add genconfig command for generating configurations ([9695c4a](https://github.com/triargos/webdav/commit/9695c4a11932a6f50b5948ba00bfe466037bdd33))
+* **config:** allow auth type specification via envíronment variable ([08275f7](https://github.com/triargos/webdav/commit/08275f77f517557a9e3aba6349e113890c8d0190))
+
+## [3.1.1-dev.1](https://github.com/triargos/webdav/compare/v3.1.0...v3.1.1-dev.1) (2024-07-23)
+
+
+### Bug Fixes
+
+* **ci:** update action ([0940685](https://github.com/triargos/webdav/commit/0940685f4d9ebab570312b7df9f6a1e95b4e9114))
+* **ci:** update action ([e96bd16](https://github.com/triargos/webdav/commit/e96bd16c2d8742aabea074aef2886f58f36b26e6))
+* **config:** remove viper ([cf3c9f2](https://github.com/triargos/webdav/commit/cf3c9f27d6fd5fc9861878c667a69772e69dc3da))
+
+# [3.1.0-dev.3](https://github.com/triargos/webdav/compare/v3.1.0-dev.2...v3.1.0-dev.3) (2024-07-17)
 
 
 ### Bug Fixes
 
 * **ci:** do not run release on pr ([dd11b78](https://github.com/triargos/webdav/commit/dd11b783f60d69bb56a56548651efafe3ff46fe1))
-* **ci:** do not run release on release ([bc86c28](https://github.com/triargos/webdav/commit/bc86c28386beb191e9a41f6ba7536c38637b13c5))
-* **ci:** use semantic release ([712cab6](https://github.com/triargos/webdav/commit/712cab6c3fb2d4f0b1422262da181e99eac62478))
-* **config:** fix default config ([b4e7855](https://github.com/triargos/webdav/commit/b4e7855294cffdafdd4281e44373106c2974b6d0))
-
-
-### Features
-
-* **authentication:** enable digest authentication ([427cb83](https://github.com/triargos/webdav/commit/427cb839a39c58e56e25a26a7dfa0687f61612a2))
 
 # [3.1.0-dev.2](https://github.com/triargos/webdav/compare/v3.1.0-dev.1...v3.1.0-dev.2) (2024-07-17)